Create nms and loss_utils files
