+++ src/transformers/models/convnext_maskrcnn/feature_extraction_convnext_maskrcnn.py
@@ -15,15 +15,17 @@
 """Feature extractor class for ConvNext Mask-RCNN."""
 
 import warnings
-from typing import Dict, List, Optional, Tuple, Union
+from typing import List, Union
 
 import numpy as np
-import torchvision
 from PIL import Image
 
 from ...feature_extraction_utils import BatchFeature, FeatureExtractionMixin
 from ...image_utils import ImageFeatureExtractionMixin
-from ...utils import TensorType, is_torch_available, logging
+from ...nms import multiclass_nms
+from ...utils import is_torch_available, logging
+
+from .modeling_convnext_maskrcnn import ConvNextMaskRCNNDeltaXYWHBBoxCoder
 
 
 if is_torch_available():
@@ -35,427 +37,10 @@ logger = logging.get_logger(__name__)
 
 ImageInput = Union[Image.Image, np.ndarray, "torch.Tensor", List[Image.Image], List[np.ndarray], List["torch.Tensor"]]
 
-ArrayType = Union[torch.Tensor, np.ndarray]
-
 BYTES_PER_FLOAT = 4
 # TODO: This memory limit may be too much or too little. It would be better to
 # determine it based on available resources.
 GPU_MEM_LIMIT = 1024**3  # 1 GB memory limit
-
-
-def nms(
-    boxes: ArrayType,
-    scores: ArrayType,
-    iou_threshold: float,
-    offset: int = 0,
-    score_threshold: float = 0,
-    max_num: int = -1,
-) -> Tuple[ArrayType, ArrayType]:
-    """Dispatch to either CPU or GPU NMS implementations.
-    Arguments:
-    The input can be either torch tensor or numpy array. GPU NMS will be used if the input is gpu tensor, otherwise CPU:
-    NMS will be used. The returned type will always be the same as inputs.
-        boxes (torch.Tensor or np.ndarray): boxes in shape (N, 4). scores (torch.Tensor or np.ndarray): scores in shape
-        (N, ). iou_threshold (float): IoU threshold for NMS. offset (int, 0 or 1): boxes' width or height is (x2 - x1 +
-        offset). score_threshold (float): score threshold for NMS. max_num (int): maximum number of boxes after NMS.
-    Returns:
-        tuple: kept dets (boxes and scores) and indice, which always have the same data type as the input.
-    Example:
-        >>> boxes = np.array([[49.1, 32.4, 51.0, 35.9], >>> [49.3, 32.9, 51.0, 35.3], >>> [49.2, 31.8, 51.0, 35.4], >>>
-        [35.1, 11.5, 39.1, 15.7], >>> [35.6, 11.8, 39.3, 14.2], >>> [35.3, 11.5, 39.9, 14.5], >>> [35.2, 11.7, 39.7,
-        15.7]], dtype=np.float32) >>> scores = np.array([0.9, 0.9, 0.5, 0.5, 0.5, 0.4, 0.3],\
-               dtype=np.float32)
-        >>> iou_threshold = 0.6 >>> dets, inds = nms(boxes, scores, iou_threshold) >>> assert len(inds) == len(dets) ==
-        3
-    """
-    assert isinstance(boxes, (torch.Tensor, np.ndarray))
-    assert isinstance(scores, (torch.Tensor, np.ndarray))
-    is_numpy = False
-    if isinstance(boxes, np.ndarray):
-        is_numpy = True
-        boxes = torch.from_numpy(boxes)
-    if isinstance(scores, np.ndarray):
-        scores = torch.from_numpy(scores)
-    assert boxes.size(1) == 4
-    assert boxes.size(0) == scores.size(0)
-    assert offset in (0, 1)
-
-    # TODO use the NMS op of mmcv: https://github.com/open-mmlab/mmcv/blob/d71d067da19d71d79e7b4d7ae967891c7bb00c05/mmcv/ops/nms.py#L28
-    # inds = NMSop.apply(boxes, scores, iou_threshold, offset, score_threshold, max_num)
-    inds = torchvision.ops.nms(boxes, scores, iou_threshold)
-    dets = torch.cat((boxes[inds], scores[inds].reshape(-1, 1)), dim=1)
-    if is_numpy:
-        dets = dets.cpu().numpy()
-        inds = inds.cpu().numpy()
-    return dets, inds
-
-
-def batched_nms(
-    boxes: torch.Tensor,
-    scores: torch.Tensor,
-    idxs: torch.Tensor,
-    nms_cfg: Optional[Dict],
-    class_agnostic: bool = False,
-) -> Tuple[torch.Tensor, torch.Tensor]:
-    r"""Performs non-maximum suppression in a batched fashion.
-    Modified from [torchvision/ops/boxes.py#L39](https://github.com/pytorch/vision/blob/
-    505cd6957711af790211896d32b40291bea1bc21/torchvision/ops/boxes.py#L39). In order to perform NMS independently per
-    class, we add an offset to all the boxes. The offset is dependent only on the class idx, and is large enough so
-    that boxes from different classes do not overlap. Note:
-        In v1.4.1 and later, `batched_nms` supports skipping the NMS and returns sorted raw results when *nms_cfg* is
-        None.
-
-    Args:
-        boxes (torch.Tensor): boxes in shape (N, 4) or (N, 5).
-        scores (torch.Tensor): scores in shape (N, ).
-        idxs (torch.Tensor): each index value correspond to a bbox cluster,
-            and NMS will not be applied between elements of different idxs, shape (N, ).
-        nms_cfg (dict | optional): Supports skipping the nms when *nms_cfg*
-            is None, otherwise it should specify nms type and other parameters like *iou_thr*. Possible keys includes
-            the following.
-            - iou_threshold (float): IoU threshold used for NMS.
-            - split_thr (float): threshold number of boxes. In some cases the number of boxes is large (e.g., 200k). To
-              avoid OOM during training, the users could set *split_thr* to a small value. If the number of boxes is
-              greater than the threshold, it will perform NMS on each group of boxes separately and sequentially.
-              Defaults to 10000.
-        class_agnostic (bool): if true, nms is class agnostic,
-            i.e. IoU thresholding happens over all boxes, regardless of the predicted class. Defaults to False.
-    Returns:
-        tuple: kept dets and indice.
-        - boxes (Tensor): Bboxes with score after nms, has shape (num_bboxes, 5). last dimension 5 arrange as (x1, y1,
-          x2, y2, score)
-        - keep (Tensor): The indices of remaining boxes in input boxes.
-    """
-    # skip nms when nms_cfg is None
-    if nms_cfg is None:
-        scores, inds = scores.sort(descending=True)
-        boxes = boxes[inds]
-        return torch.cat([boxes, scores[:, None]], -1), inds
-
-    nms_cfg_ = nms_cfg.copy()
-    class_agnostic = nms_cfg_.pop("class_agnostic", class_agnostic)
-    if class_agnostic:
-        boxes_for_nms = boxes
-    else:
-        # When using rotated boxes, only apply offsets on center.
-        if boxes.size(-1) == 5:
-            # Strictly, the maximum coordinates of the rotating box
-            # (x,y,w,h,a) should be calculated by polygon coordinates.
-            # But the conversion from rotated box to polygon will
-            # slow down the speed.
-            # So we use max(x,y) + max(w,h) as max coordinate
-            # which is larger than polygon max coordinate
-            # max(x1, y1, x2, y2,x3, y3, x4, y4)
-            max_coordinate = boxes[..., :2].max() + boxes[..., 2:4].max()
-            offsets = idxs.to(boxes) * (max_coordinate + torch.tensor(1).to(boxes))
-            boxes_ctr_for_nms = boxes[..., :2] + offsets[:, None]
-            boxes_for_nms = torch.cat([boxes_ctr_for_nms, boxes[..., 2:5]], dim=-1)
-        else:
-            max_coordinate = boxes.max()
-            offsets = idxs.to(boxes) * (max_coordinate + torch.tensor(1).to(boxes))
-            boxes_for_nms = boxes + offsets[:, None]
-
-    nms_type = nms_cfg_.pop("type", "nms")
-    nms_op = eval(nms_type)
-
-    split_thr = nms_cfg_.pop("split_thr", 10000)
-    # Won't split to multiple nms nodes when exporting to onnx
-    if boxes_for_nms.shape[0] < split_thr or torch.onnx.is_in_onnx_export():
-        dets, keep = nms_op(boxes_for_nms, scores, **nms_cfg_)
-        boxes = boxes[keep]
-
-        # This assumes `dets` has arbitrary dimensions where
-        # the last dimension is score.
-        # Currently it supports bounding boxes [x1, y1, x2, y2, score] or
-        # rotated boxes [cx, cy, w, h, angle_radian, score].
-
-        scores = dets[:, -1]
-    else:
-        max_num = nms_cfg_.pop("max_num", -1)
-        total_mask = scores.new_zeros(scores.size(), dtype=torch.bool)
-        # Some type of nms would reweight the score, such as SoftNMS
-        scores_after_nms = scores.new_zeros(scores.size())
-        for id in torch.unique(idxs):
-            mask = (idxs == id).nonzero(as_tuple=False).view(-1)
-            dets, keep = nms_op(boxes_for_nms[mask], scores[mask], **nms_cfg_)
-            total_mask[mask[keep]] = True
-            scores_after_nms[mask[keep]] = dets[:, -1]
-        keep = total_mask.nonzero(as_tuple=False).view(-1)
-
-        scores, inds = scores_after_nms[keep].sort(descending=True)
-        keep = keep[inds]
-        boxes = boxes[keep]
-
-        if max_num > 0:
-            keep = keep[:max_num]
-            boxes = boxes[:max_num]
-            scores = scores[:max_num]
-
-    boxes = torch.cat([boxes, scores[:, None]], -1)
-    return boxes, keep
-
-
-def multiclass_nms(multi_bboxes, multi_scores, score_thr, nms_cfg, max_num=-1, score_factors=None, return_inds=False):
-    """NMS for multi-class bboxes.
-    Args:
-        multi_bboxes (Tensor): shape (n, #class*4) or (n, 4)
-        multi_scores (Tensor): shape (n, #class), where the last column
-            contains scores of the background class, but this will be ignored.
-        score_thr (float): bbox threshold, bboxes with scores lower than it
-            will not be considered.
-        nms_thr (float): NMS IoU threshold
-        max_num (int, optional): if there are more than max_num bboxes after
-            NMS, only top max_num will be kept. Default to -1.
-        score_factors (Tensor, optional): The factors multiplied to scores
-            before applying NMS. Default to None.
-        return_inds (bool, optional): Whether return the indices of kept
-            bboxes. Default to False.
-    Returns:
-        tuple: (dets, labels, indices (optional)), tensors of shape (k, 5),
-            (k), and (k). Dets are boxes with scores. Labels are 0-based.
-    """
-    num_classes = multi_scores.size(1) - 1
-    # exclude background category
-    if multi_bboxes.shape[1] > 4:
-        bboxes = multi_bboxes.view(multi_scores.size(0), -1, 4)
-    else:
-        bboxes = multi_bboxes[:, None].expand(multi_scores.size(0), num_classes, 4)
-
-    scores = multi_scores[:, :-1]
-
-    labels = torch.arange(num_classes, dtype=torch.long, device=scores.device)
-    labels = labels.view(1, -1).expand_as(scores)
-
-    bboxes = bboxes.reshape(-1, 4)
-    scores = scores.reshape(-1)
-    labels = labels.reshape(-1)
-
-    if not torch.onnx.is_in_onnx_export():
-        # NonZero not supported  in TensorRT
-        # remove low scoring boxes
-        valid_mask = scores > score_thr
-    # multiply score_factor after threshold to preserve more bboxes, improve
-    # mAP by 1% for YOLOv3
-    if score_factors is not None:
-        # expand the shape to match original shape of score
-        score_factors = score_factors.view(-1, 1).expand(multi_scores.size(0), num_classes)
-        score_factors = score_factors.reshape(-1)
-        scores = scores * score_factors
-
-    if not torch.onnx.is_in_onnx_export():
-        # NonZero not supported  in TensorRT
-        inds = valid_mask.nonzero(as_tuple=False).squeeze(1)
-        bboxes, scores, labels = bboxes[inds], scores[inds], labels[inds]
-    else:
-        # TensorRT NMS plugin has invalid output filled with -1
-        # add dummy data to make detection output correct.
-        bboxes = torch.cat([bboxes, bboxes.new_zeros(1, 4)], dim=0)
-        scores = torch.cat([scores, scores.new_zeros(1)], dim=0)
-        labels = torch.cat([labels, labels.new_zeros(1)], dim=0)
-
-    if bboxes.numel() == 0:
-        if torch.onnx.is_in_onnx_export():
-            raise RuntimeError("[ONNX Error] Can not record NMS as it has not been executed this time")
-        dets = torch.cat([bboxes, scores[:, None]], -1)
-        if return_inds:
-            return dets, labels, inds
-        else:
-            return dets, labels
-
-    dets, keep = batched_nms(bboxes, scores, labels, nms_cfg)
-
-    if max_num > 0:
-        dets = dets[:max_num]
-        keep = keep[:max_num]
-
-    if return_inds:
-        return dets, labels[keep], inds[keep]
-    else:
-        return dets, labels[keep]
-
-
-class ConvNextMaskRCNNDeltaXYWHBBoxCoder(nn.Module):
-    """Delta XYWH BBox coder.
-    Following the practice in [R-CNN](https://arxiv.org/abs/1311.2524), this coder encodes bbox (x1, y1, x2, y2) into
-    delta (dx, dy, dw, dh) and decodes delta (dx, dy, dw, dh) back to original bbox (x1, y1, x2, y2).
-
-    Args:
-        target_means (Sequence[float]): Denormalizing means of target for
-            delta coordinates
-        target_stds (Sequence[float]): Denormalizing standard deviation of
-            target for delta coordinates
-        clip_border (bool, optional): Whether clip the objects outside the
-            border of the image. Defaults to True.
-        add_ctr_clamp (bool): Whether to add center clamp, when added, the
-            predicted box is clamped is its center is too far away from the original anchor's center. Only used by
-            YOLOF. Default False.
-        ctr_clamp (int): the maximum pixel shift to clamp. Only used by YOLOF.
-            Default 32.
-    """
-
-    def __init__(
-        self,
-        target_means=(0.0, 0.0, 0.0, 0.0),
-        target_stds=(1.0, 1.0, 1.0, 1.0),
-        clip_border=True,
-        add_ctr_clamp=False,
-        ctr_clamp=32,
-    ):
-        super().__init__()
-        self.means = target_means
-        self.stds = target_stds
-        self.clip_border = clip_border
-        self.add_ctr_clamp = add_ctr_clamp
-        self.ctr_clamp = ctr_clamp
-
-    def decode(self, bboxes, pred_bboxes, max_shape=None, wh_ratio_clip=16 / 1000):
-        """Apply transformation `pred_bboxes` to `boxes`.
-
-        Args:
-            bboxes (torch.Tensor):
-                Basic boxes. Shape (B, N, 4) or (N, 4)
-            pred_bboxes (Tensor):
-                Encoded offsets with respect to each roi. Has shape (B, N, num_classes * 4) or (B, N, 4) or (N,
-                num_classes * 4) or (N, 4). Note N = num_anchors * W * H when rois is a grid of anchors.Offset encoding
-                follows [1]_.
-            max_shape (Sequence[int] or torch.Tensor or Sequence[
-               Sequence[int]],optional): Maximum bounds for boxes, specifies (H, W, C) or (H, W). If bboxes shape is
-               (B, N, 4), then the max_shape should be a Sequence[Sequence[int]] and the length of max_shape should
-               also be B.
-            wh_ratio_clip (float, optional): The allowed ratio between
-                width and height.
-
-        Returns:
-            torch.Tensor: Decoded boxes.
-        """
-
-        assert pred_bboxes.size(0) == bboxes.size(0)
-        if pred_bboxes.ndim == 3:
-            assert pred_bboxes.size(1) == bboxes.size(1)
-
-        if pred_bboxes.ndim == 2 and not torch.onnx.is_in_onnx_export():
-            # single image decode
-            decoded_bboxes = delta2bbox(
-                bboxes,
-                pred_bboxes,
-                self.means,
-                self.stds,
-                max_shape,
-                wh_ratio_clip,
-                self.clip_border,
-                self.add_ctr_clamp,
-                self.ctr_clamp,
-            )
-        else:
-            if pred_bboxes.ndim == 3 and not torch.onnx.is_in_onnx_export():
-                # TODO: remove this deprecation
-                warnings.warn(
-                    "DeprecationWarning: onnx_delta2bbox is deprecated "
-                    "in the case of batch decoding and non-ONNX, "
-                    "please use “delta2bbox” instead. In order to improve "
-                    "the decoding speed, the batch function will no "
-                    "longer be supported. "
-                )
-            raise NotImplementedError("ONNX is not yet supported")
-            # decoded_bboxes = onnx_delta2bbox(
-            #     bboxes,
-            #     pred_bboxes,
-            #     self.means,
-            #     self.stds,
-            #     max_shape,
-            #     wh_ratio_clip,
-            #     self.clip_border,
-            #     self.add_ctr_clamp,
-            #     self.ctr_clamp,
-            # )
-
-        return decoded_bboxes
-
-
-def delta2bbox(
-    rois,
-    deltas,
-    means=(0.0, 0.0, 0.0, 0.0),
-    stds=(1.0, 1.0, 1.0, 1.0),
-    max_shape=None,
-    wh_ratio_clip=16 / 1000,
-    clip_border=True,
-    add_ctr_clamp=False,
-    ctr_clamp=32,
-):
-    """Apply deltas to shift/scale base boxes.
-    Args:
-    Typically the rois are anchor or proposed bounding boxes and the deltas are network outputs used to shift/scale
-    those boxes. This is the inverse function of [`bbox2delta`].
-        rois (Tensor): Boxes to be transformed. Has shape (N, 4). deltas (Tensor): Encoded offsets relative to each
-        roi.
-            Has shape (N, num_classes * 4) or (N, 4). Note N = num_base_anchors * W * H, when rois is a grid of
-            anchors. Offset encoding follows [1]_.
-        means (Sequence[float]): Denormalizing means for delta coordinates.
-            Default (0., 0., 0., 0.).
-        stds (Sequence[float]): Denormalizing standard deviation for delta
-            coordinates. Default (1., 1., 1., 1.).
-        max_shape (tuple[int, int]): Maximum bounds for boxes, specifies
-           (H, W). Default None.
-        wh_ratio_clip (float): Maximum aspect ratio for boxes. Default
-            16 / 1000.
-        clip_border (bool, optional): Whether clip the objects outside the
-            border of the image. Default True.
-        add_ctr_clamp (bool): Whether to add center clamp. When set to True,
-            the center of the prediction bounding box will be clamped to avoid being too far away from the center of
-            the anchor. Only used by YOLOF. Default False.
-        ctr_clamp (int): the maximum pixel shift to clamp. Only used by YOLOF.
-            Default 32.
-    Returns:
-        Tensor: Boxes with shape (N, num_classes * 4) or (N, 4), where 4
-           represent tl_x, tl_y, br_x, br_y.
-    References:
-        .. [1] https://arxiv.org/abs/1311.2524
-    Example:
-        >>> rois = torch.Tensor([[ 0., 0., 1., 1.], >>> [ 0., 0., 1., 1.], >>> [ 0., 0., 1., 1.], >>> [ 5., 5., 5.,
-        5.]]) >>> deltas = torch.Tensor([[ 0., 0., 0., 0.], >>> [ 1., 1., 1., 1.], >>> [ 0., 0., 2., -1.], >>> [ 0.7,
-        -1.9, -0.5, 0.3]]) >>> delta2bbox(rois, deltas, max_shape=(32, 32, 3)) tensor([[0.0000, 0.0000, 1.0000,
-        1.0000],
-                [0.1409, 0.1409, 2.8591, 2.8591], [0.0000, 0.3161, 4.1945, 0.6839], [5.0000, 5.0000, 5.0000, 5.0000]])
-    """
-    num_bboxes, num_classes = deltas.size(0), deltas.size(1) // 4
-    if num_bboxes == 0:
-        return deltas
-
-    deltas = deltas.reshape(-1, 4)
-
-    means = deltas.new_tensor(means).view(1, -1)
-    stds = deltas.new_tensor(stds).view(1, -1)
-    denorm_deltas = deltas * stds + means
-
-    dxy = denorm_deltas[:, :2]
-    dwh = denorm_deltas[:, 2:]
-
-    # Compute width/height of each roi
-    rois_ = rois.repeat(1, num_classes).reshape(-1, 4)
-    pxy = (rois_[:, :2] + rois_[:, 2:]) * 0.5
-    pwh = rois_[:, 2:] - rois_[:, :2]
-
-    dxy_wh = pwh * dxy
-
-    max_ratio = np.abs(np.log(wh_ratio_clip))
-    if add_ctr_clamp:
-        dxy_wh = torch.clamp(dxy_wh, max=ctr_clamp, min=-ctr_clamp)
-        dwh = torch.clamp(dwh, max=max_ratio)
-    else:
-        dwh = dwh.clamp(min=-max_ratio, max=max_ratio)
-
-    gxy = pxy + dxy_wh
-    gwh = pwh * dwh.exp()
-    x1y1 = gxy - (gwh * 0.5)
-    x2y2 = gxy + (gwh * 0.5)
-    bboxes = torch.cat([x1y1, x2y2], dim=-1)
-    if clip_border and max_shape is not None:
-        max_shape = max_shape[-2:]
-        bboxes[..., 0::2].clamp_(min=0, max=max_shape[1])
-        bboxes[..., 1::2].clamp_(min=0, max=max_shape[0])
-    bboxes = bboxes.reshape(num_bboxes, -1)
-    return bboxes
 
 
 def _do_paste_mask(masks, boxes, img_h, img_w, skip_empty=True):
@@ -772,9 +357,9 @@ class ConvNextMaskRCNNFeatureExtractor(FeatureExtractionMixin, ImageFeatureExtra
         img_shapes = []
         for target_size, scale_factor in zip(target_sizes, scale_factors):
             height, width = target_size[-2:]
-            img_shape = (3, height*scale_factor[1], width*scale_factor[0])
+            img_shape = (3, height * scale_factor[1], width * scale_factor[0])
             img_shapes.append(img_shape)
-        
+
         # apply bbox post-processing to each image individually
         det_bboxes = []
         det_labels = []
@@ -794,7 +379,7 @@ class ConvNextMaskRCNNFeatureExtractor(FeatureExtractionMixin, ImageFeatureExtra
                     bbox_pred[i],
                     img_shapes[i],
                     scale_factors[i],
-                    rescale=True, # we rescale by default
+                    rescale=True,  # we rescale by default
                     cfg=self.test_cfg,
                 )
             det_bboxes.append(det_bbox)
@@ -817,7 +402,11 @@ class ConvNextMaskRCNNFeatureExtractor(FeatureExtractionMixin, ImageFeatureExtra
         return results
 
     def post_process_instance_segmentation(
-        self, object_detection_results, mask_pred, target_sizes=None, scale_factors=None,
+        self,
+        object_detection_results,
+        mask_pred,
+        target_sizes=None,
+        scale_factors=None,
     ):
         det_bboxes = [result["boxes"] for result in object_detection_results]
         det_labels = [result["labels"] for result in object_detection_results]
@@ -828,7 +417,7 @@ class ConvNextMaskRCNNFeatureExtractor(FeatureExtractionMixin, ImageFeatureExtra
 
         num_imgs = len(object_detection_results)
 
-        rescale = True # we rescale by default
+        rescale = True  # we rescale by default
         scale_factors = [torch.from_numpy(scale_factor).to(det_bboxes[0].device) for scale_factor in scale_factors]
         _bboxes = [
             # det_bboxes[i][:, :4] * scale_factors[i] if rescale else det_bboxes[i][:, :4]
